【viewer】viewer C化bug修复
# message auto-generated for no-merge-commit merge:
merge viewer_ccc into master

【viewer】viewer C化bug修复

Created-by: msprofDeveloper

Match-id-30f959b74e0b7fa513237c6bc3fecbde0b1f7e57

+++ analysis/csrc/domain/entities/viewer_data/system/include/soc_bandwidth_data.h
@@ -18,7 +18,7 @@
 namespace Analysis {
 namespace Domain {
 struct SocBandwidthData {
-    uint16_t deviceId = UINT32_MAX;
+    uint16_t deviceId = UINT16_MAX;
     uint32_t l2_buffer_bw_level = 0;
     uint32_t mata_bw_level = 0;
     uint64_t timestamp = 0;
